Show Ctrl+/ instead of ⌘K on all platforms but Mac
